feat: refactor modal and input components, adding new files and functionalities
