+++ src/ui/global/widgets/components/input/index.tsx
@@ -1,0 +1,3 @@
+import { InputView } from './input-view'
+
+export const Input = InputView
