+++ src/ui/global/widgets/components/modal/modal-ref.ts
@@ -1,0 +1,4 @@
+export interface ModalRef {
+  open: () => void
+  close: () => void
+}
